// src/components/IconSvg/type.ts
export type IconSvgType =
  | 'twitter'
  | 'apple'
  | 'calendar'
  | 'camera'
  | 'check'
  | 'edit'
  | 'facebook'
  | 'google'
  | 'idk'
  | 'left_arrow'
  | 'lock'
  | 'logo'
  | 'map'
  | 'noti'
  | 'refresh'
  | 'scan'
  | 'search'
  | 'setting'
  | 'user_plus'
  | 'user'
  | 'kyc'
  | 'location_point'
  | 'check_white'
  | 'card'
  | 'cmnd'
  | 'selfie'
  | 'green_check'
  | 'check_green'
  | 'red_x'
  | 'fail'
  | 'privacy_tip'
  | 'item_marker'
  | 'reward'
  | 'wallet'
  | 'preference'
  | 'copy'
  | 'right_v'
  | 'phone';